+++ sorna/agent/gpu.py
@@ -18,7 +18,8 @@ async def prepare_nvidia(docker, numa_node):
 
     volumes = await docker.volumes.list()
     existing_volumes = set(vol['Name'] for vol in volumes['Volumes'])
-    required_volumes = set(vol.split(':')[0] for vol in nvidia_params['Volumes'])
+    required_volumes = set(vol.split(':')[0]
+                           for vol in nvidia_params['Volumes'])
     missing_volumes = required_volumes - existing_volumes
     binds = []
     for vol_name in missing_volumes:
@@ -26,7 +27,10 @@ async def prepare_nvidia(docker, numa_node):
             if vol_param.startswith(vol_name + ':'):
                 _, _, permission = vol_param.split(':')
                 driver = nvidia_params['VolumeDriver']
-                await docker.volumes.create({'Name': vol_name, 'Driver': driver})
+                await docker.volumes.create({
+                    'Name': vol_name,
+                    'Driver': driver,
+                })
     for vol_name in required_volumes:
         for vol_param in nvidia_params['Volumes']:
             if vol_param.startswith(vol_name + ':'):